Clean up code base
