+++ TTT.sln
@@ -29,6 +29,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TTT.Logs", "mod\TTT.Logs\TT
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TTT.Player", "mod\TTT.Player\TTT.Player.csproj", "{D9327D60-67EE-4AF3-BD99-8A95634600B5}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TTT.Manager", "mod\TTT.Manager\TTT.Manager.csproj", "{C3C26EF9-0E5E-4AAF-AE00-EFF4BB129FDB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -78,6 +80,10 @@ Global
 		{D9327D60-67EE-4AF3-BD99-8A95634600B5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D9327D60-67EE-4AF3-BD99-8A95634600B5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D9327D60-67EE-4AF3-BD99-8A95634600B5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C3C26EF9-0E5E-4AAF-AE00-EFF4BB129FDB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C3C26EF9-0E5E-4AAF-AE00-EFF4BB129FDB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C3C26EF9-0E5E-4AAF-AE00-EFF4BB129FDB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C3C26EF9-0E5E-4AAF-AE00-EFF4BB129FDB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{A917B0F5-762F-4C4A-9F50-02BA4A0C6B63} = {C7CCF187-9F99-4091-B092-6397C35BE9F1}
@@ -90,5 +96,6 @@ Global
 		{C6EE882B-E597-45DB-85FF-A521DC7A0AAE} = {C84CEB1F-0132-43EF-9A5F-12C7AA08B8E7}
 		{E76CD078-0F1B-4EE8-9A19-EE93BAE3FA6B} = {14FF3D75-F9CF-4A32-9F59-83765EA68874}
 		{D9327D60-67EE-4AF3-BD99-8A95634600B5} = {14FF3D75-F9CF-4A32-9F59-83765EA68874}
+		{C3C26EF9-0E5E-4AAF-AE00-EFF4BB129FDB} = {14FF3D75-F9CF-4A32-9F59-83765EA68874}
 	EndGlobalSection
 EndGlobal
