+++ mod/TTT.Manager/InfoManager.cs
@@ -6,6 +6,7 @@ using CounterStrikeSharp.API.Modules.Admin;
 using CounterStrikeSharp.API.Modules.Timers;
 using CounterStrikeSharp.API.Modules.Utils;
 using TTT.Player;
+using TTT.Public.Behaviors;
 using TTT.Public.Extensions;
 using TTT.Public.Formatting;
 using TTT.Public.Mod.Role;
@@ -14,21 +15,25 @@ using TTT.Public.Player;
 using TTT.Round;
 using Vector = CounterStrikeSharp.API.Modules.Utils.Vector;
 
-namespace TTT.Roles;
+namespace TTT.Manager;
 
-public class InfoManager
+public class InfoManager : IPluginBehavior
 {
     private readonly Dictionary<CCSPlayerController, Tuple<CCSPlayerController, Role>> _playerLookAtRole = new();
     private readonly IPlayerService _playerService;
     private readonly IRoundService _manager;
     private readonly Dictionary<CCSPlayerController, Tuple<string, Role>> _spectatorLookAtRole = new();
 
-    public InfoManager(IPlayerService playerService, IRoundService manager, BasePlugin plugin)
+    public InfoManager(IPlayerService playerService, IRoundService manager)
     {
         _playerService = playerService;
         _manager = manager;
+    }
+    
+    public void Start(BasePlugin plugin)
+    {
         plugin.RegisterListener<Listeners.OnTick>(OnTick);
-        plugin.AddTimer(0.3f, OnTickAll, TimerFlags.REPEAT);
+        plugin.AddTimer(0.1f, OnTickAll, TimerFlags.REPEAT);
 
         plugin.RegisterEventHandler<EventSpecTargetUpdated>(OnPlayerSpectateChange);
     }
@@ -85,6 +90,8 @@ public class InfoManager
                 
                 continue;
             }
+            
+            if (!player.PawnIsAlive) continue;
 
             if (!_playerLookAtRole.TryGetValue(player, out var value))
             {
