+++ mod/TTT.Roles/RoleServiceExtension.cs
@@ -9,8 +9,7 @@ public static class RoleServiceExtension
 {
     public static void AddTTTRoles(this IServiceCollection collection)
     {
-        collection.AddPluginBehavior<IRoleService, RoleManager>();
-        collection.AddPluginBehavior<ChatManager>();
+        collection.AddPluginBehavior<IRoleService, RoleBehavior>();
         //collection.AddPluginBehavior<RolesCommand>();
     }
 }
