+++ public/TTT.Public/Mod/Manager/IMuteService.cs
@@ -1,0 +1,11 @@
+﻿using CounterStrikeSharp.API.Core;
+
+namespace TTT.Public.Mod.Manager;
+
+public interface IMuteService
+{
+    void Mute(CCSPlayerController player);
+    void UnMute(CCSPlayerController player);
+    void UnMuteAll();
+    void MuteAll();
+}
