+++ src/TTT/TTT.csproj
@@ -4,11 +4,16 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
+  
   <ItemGroup>
     <ProjectReference Include="..\..\public\TTT.Public\TTT.Public.csproj" />
     <ProjectReference Include="..\..\mod\TTT.Detective\TTT.Detective.csproj" />
     <ProjectReference Include="..\..\mod\TTT.Roles\TTT.Roles.csproj" />
+    <ProjectReference Include="..\..\mod\TTT.Player\TTT.Player.csproj" />
+    <ProjectReference Include="..\..\mod\TTT.Manager\TTT.Manager.csproj" />
+    <ProjectReference Include="..\..\mod\TTT.Logs\TTT.Logs.csproj" />
   </ItemGroup>
+  
   <PropertyGroup>
     <!--Publish Configuration-->
     <PublishBaseDirectory>$(MSBuildThisFileDirectory)/../../build</PublishBaseDirectory>
@@ -17,6 +22,7 @@
     <PublishSelfContained>false</PublishSelfContained>
     <PublishDocumentationFile>true</PublishDocumentationFile>
   </PropertyGroup>
+  
   <Target Name="ZipOutputPath" AfterTargets="Publish">
     <ZipDirectory Overwrite="true" SourceDirectory="$(PublishDir)\" DestinationFile="$(PublishBaseDirectory)\TTT.zip" />
   </Target>
